backend/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
Django==5.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
idna==3.10
mysqlclient==2.2.7
PyJWT==2.10.1
requests==2.32.3
sentry-sdk==2.10.0
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
sentry-sdk[django]
django-allauth==0.54.0
dj-rest-auth==4.0.1
django-oauth-toolkit==2.3.0
social-auth-app-django==5.2.0
django-redis==5.4.0
redis==5.0.1
django-filter==23.5
openai>=1.12.0  # For AI-based test result extraction
celery>=5.3.0
django-celery-beat>=2.5.0
django-redis>=5.4.0

